Changes
- Renamed DashboardNewDeploy to NewDeployNotAvailabe
- Renamed Dashboard404 to Project404 and moved to shared
- Reworked NewDeployNotAvailable to use suggestions
- Removed Feedback Modal from v1
