+++ src/components/v2/V2Dashboard/Gateway.tsx
@@ -4,7 +4,7 @@ import { useParams } from 'react-router-dom'
 import Loading from 'components/shared/Loading'
 
 import V2Dashboard from '.'
-import Dashboard404 from './Dashboard404'
+import Project404 from '../../shared/Project404'
 
 const parseProjectIdParameter = (projectIdParameter?: string) => {
   try {
@@ -28,8 +28,7 @@ export default function V2DashboardGateway() {
 
   const projectId =
     projectIdForName ?? parseProjectIdParameter(projectIdParameter)
-  if (!projectId)
-    return <Dashboard404 projectId={projectId} ensName={ensName} />
+  if (!projectId) return <Project404 projectId={projectId} ensName={ensName} />
 
   return <V2Dashboard projectId={projectId} />
 }
